Moved openaerostruct to aerodynamics folder
