+++ openconcept/aerodynamics/__init__.py
@@ -1,1 +1,2 @@
 from .aerodynamics import PolarDrag, StallSpeed, Lift
+from .openaerostruct import OASDragPolar, OASAerostructDragPolar
