+++ openconcept/aerodynamics/openaerostruct/tests/test_drag_polar.py
@@ -7,7 +7,7 @@ import openmdao.api as om
 try:
     from openaerostruct.geometry.geometry_group import Geometry
     from openaerostruct.aerodynamics.aero_groups import AeroPoint
-    from openconcept.analysis.openaerostruct.drag_polar import *
+    from openconcept.aerodynamics.openaerostruct.drag_polar import *
 
     OAS_installed = True
 except:
