+++ openconcept/examples/B738_VLM_drag.py
@@ -20,7 +20,7 @@ import numpy as np
 import openmdao.api as om
 import openconcept.api as oc
 # imports for the airplane model itself
-from openconcept.analysis.openaerostruct.drag_polar import OASDragPolar
+from openconcept.aerodynamics import OASDragPolar
 from openconcept.examples.aircraft_data.B738 import data as acdata
 from openconcept.analysis.performance.mission_profiles import MissionWithReserve
 from openconcept.components.cfm56 import CFM56
